Renamed FileMetaData to FileMetaDataDTO
